--- fastapi_forge/__init__.py
@@ -1,4 +1,0 @@
-from .cli import main
-
-
-main()
