--- fastapi_forge/main.py
@@ -1,4 +1,0 @@
-from .cli import main
-
-if __name__ in {"__main__", "__mp_main__"}:
-    main()
