+++ pyproject.toml
@@ -1,0 +1,19 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "fastapi-forge"
+description = "Generate FastAPI projects based on database schema."
+readme = "README.md"
+requires-python = ">=3.12"
+dynamic = ["version"]
+dependencies = ["click", "nicegui", "cookiecutter"]
+
+[project.scripts]
+fastapi-forge = "fastapi_forge.__main__:main"
+
+[tool.ruff]
+exclude = ["template"]
+line-length = 88
+indent-width = 4
