--- requirements.txt
@@ -1,87 +1,0 @@
-aiofiles==24.1.0
-aiohappyeyeballs==2.4.4
-aiohttp==3.11.11
-aiosignal==1.3.2
-annotated-types==0.7.0
-anyio==4.8.0
-arrow==1.3.0
-attrs==25.1.0
-bidict==0.23.1
-binaryornot==0.4.4
-certifi==2024.12.14
-cffi==1.17.1
-chardet==5.2.0
-charset-normalizer==3.4.1
-click==8.1.8
-colorama==0.4.6
-cookiecutter==2.6.0
-cryptography==44.0.0
-docutils==0.21.2
-fastapi==0.115.6
-frozenlist==1.5.0
-h11==0.14.0
-httpcore==1.0.7
-httptools==0.6.4
-httpx==0.28.1
-idna==3.10
-ifaddr==0.2.0
-iniconfig==2.0.0
-itsdangerous==2.2.0
-jaraco.classes==3.4.0
-jaraco.context==6.0.1
-jaraco.functools==4.1.0
-jeepney==0.8.0
-Jinja2==3.1.5
-keyring==25.6.0
-markdown-it-py==3.0.0
-markdown2==2.5.3
-MarkupSafe==3.0.2
-mdurl==0.1.2
-more-itertools==10.5.0
-multidict==6.1.0
-nh3==0.2.20
-nicegui==2.10.1
-orjson==3.10.15
-packaging==24.2
-pkginfo==1.12.0
-pluggy==1.5.0
-propcache==0.2.1
-pscript==0.7.7
-pycparser==2.22
-pydantic==2.10.5
-pydantic_core==2.27.2
-Pygments==2.19.1
-pytest==8.3.4
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-python-engineio==4.11.2
-python-multipart==0.0.20
-python-slugify==8.0.4
-python-socketio==5.12.1
-pywin32-ctypes==0.2.3
-PyYAML==6.0.2
-readme_renderer==44.0
-requests==2.32.3
-requests-toolbelt==1.0.0
-rfc3986==2.0.0
-rich==13.9.4
-ruff==0.9.1
-SecretStorage==3.3.3
-setuptools==75.8.0
-simple-websocket==1.1.0
-six==1.17.0
-sniffio==1.3.1
-starlette==0.41.3
-text-unidecode==1.3
-twine==6.0.1
-types-python-dateutil==2.9.0.20241206
-typing_extensions==4.12.2
-urllib3==2.3.0
-uvicorn==0.34.0
-uvloop==0.21.0
-vbuild==0.8.2
-watchfiles==1.0.4
-websockets==14.2
-wheel==0.45.1
-wsproto==1.2.0
-yarl==1.18.3
