--- ruff.toml
@@ -1,1 +1,0 @@
-exclude = ["template"]
